Clarify usage of various terms from other documents.
This patch addresses the first ~half of the comments in
https://github.com/w3c/webappsec-secure-contexts/issues/33, focusing on clarity
around the usage of 'environment settings objects' and 'global object', as well
as various and sundry ways of getting to both.

Additionally, it moves to new terminology around origins, and cleans up the
document's references a little bit. Not done, but a good first step.