Remove reference to "filesystem:" scheme from a note.
This is proprietary feature and Chromium does not plan to push for
adoption [1]. The HTML spec treats the origin of an URL with such a
scheme as opaque [2]. This only seems to be mentioned in a
(non-normative) note, so it is safe to just remove it to avoid any
confusion. Also this does not really match the "Therefore, blobs..."
in the next sentence.

[1] https://groups.google.com/a/chromium.org/g/blink-dev/c/nrpl_ewkmaQ/m/mmpummdqBAAJ
[2] https://url.spec.whatwg.org/#concept-url-origin